Add --environment option to db/app commands
Most db and app commands accepting a handle needs to accept --account
(ahem --environment) as an option, since nowadays an app or db handle is
no longer globally unique.

Git remotes are also parsed so that if an account is in the git remote,
that account will be the one that is assumed by default.

By popular demand, --account has also been renamed to --environment.