Add size option when scaling app
This commit adds an optional --size parameter, which can be used to
set a memory limit on the service that is being scaled, as in the
dashboard. If not specified, the service will either whatever default
it previously had, or whatever Sweetness has set as default. The valid
values for size are the same as in the api; 512, 1204, 2048 etc.

This commit also adds Webmock to the specs, since there are some
inadvertent calls that are being made to our api.