Check for token in APTIBLE_ACCESS_TOKEN
For unattended operation, it's easier to pass a token via the
environment rather than generate a JSON file in HOME and remember to
delete it, or use `aptible login`.

This change lets the aptible CLI look for a token in the
APTIBLE_ACCESS_TOKEN environment variable. The variable takes precedence
over the Aptible configuration file to allow for predictable behavior
when running on a host where there is also a configuration file.

In particular, this feature is useful to use the aptible cli in
integration tests.