Git remote: either guess everything, or nothing.
We've had several tickets recently that came up due to customers not
realizing (and who could blame them) that even if they provide an app
handle and no environment, we're overriding their choice of environment
based on the current git remote.

This patch updates the aptible CLI to either guess both the app handle
and the environment handle from the current git remote (and explain what
it did), or guess neither if CLI arguments were provided.