Explain environment guesses to the user
We've had several tickets recently that came up due to customers not
realizing (and who could blame them) that even if they provide an app
handle and no environment, we're overriding their choice of environment
based on the current git remote.

This patch updates the aptible CLI to explain the decisions it makes,
and specifically to explain what environment it looked in (if it did
scope its search to a single environment), and where that came from.