aptible login: Let users provide a token lifetime
This allows users to provide a token lifetime when logging in (and lets
them know about the option). The default lifetime is now 1 week (12
hours when 2FA is enabled), and users can tweak it as they please
(provided it's accepted by auth.aptible.com, of course).

This also updates our aptible-auth dependency to pull in an updated
oauth2-aptible that understands `expires_at` as it's returned by our
API.