Normalize SSH args, add KeepAlive, TCPKeepAlive
This normalizes our SSH args into a single method to compose them for
all our requests, and adds the KeepAlive and TCPKeepAlive options, as
well as tunes down the log level (this will prevent the message
indicating a new key from showing).

This also stops sending the user's environment wholesale to our servers
(which of course we never used!), and instead relies on each method to
send just what it needs.