db:tunnel: ignore Errno::ESRCH when kill'ing ssh
On Linux, we can send SIGHUP to a dead process without issues, but on
Windows (including "Bash for Windows"), this happens to fail. Similarly,
waiting for the process fails too.

In practice, it doesn't really matter if these steps fail (since the
intention was to stop the process anyway), so we just ignore it.

This is particularly useful since we'll call `stop` if we get an
exception when running `ssh`, and if `stop` fails we'll never get to see
the original exception.