Dynamic Uno Support
This updates the Aptible CLI to use Dynamic Uno (as opposed to Dumptruck
or Broadwayjoe) when accessing Aptible over SSH (e.g. for aptible ssh,
logs, etc.).

The main difference here is that the process to connect is now to:

- Create an Operation (this was something we used to do in a variety of
  places already of course, but now it's also necessary for e.g.
  `aptible ssh`, `aptible db:tunnel` and `aptible ps`).
- Create an associated SshPortalConnection.
- Log in using the credentials for that connection.

The test coverage for this exists for the most part in
aptible-integration (and in aptible-api for the bits that manage the
credentials we get when creating the portal connection), but I'm open to
ideas if there are ways to test this properly here (right now, I'm a
little worried that any specs will essentially end up a stub-fest).