Add aptible operation:cancel OPERATION_ID
This adds the ability to cancel an operation (more specifically: to
request cancellation - only operations that support it will care for
that flag).

Ideally, I'd like this to eventually attach to the operation so you can
see it complete (and to have a matching `aptible operation:attach`
command), but this wouldn't be terribly useful right now since our
operation logging mechanism does not support having multiple clients
connected at the same time (this is perhaps something we'd fix
indirectly by sending operation logs to a Log Drain, where we have all
the infrastructure in place to deal with this already in tail drains).