Use our own SSH configuration file as well
More recent versions of SSH (the ones we package with the toolbelt to
support UNIX socket forwarding) are stricter about the permissions of
the SSH configuration file, so they'll fail if the owner is wrong.

Unfortunately, that tends to happen on e.g. Travis. So, to work around
this, we use our own SSH configuration file. We already override
practically everything that matters anyway, so this doesn't make a huge
difference, but it does mean users that want to customize how we connect
over SSH will need to update a different configuration file (which is
probably for the best).