Fix Windows exit tunnel
It looks like we need to apply more force when shutting down OpenSSH on
Windows, at least when using the Windows fork
(https://github.com/PowerShell/Win32-OpenSSH /
https://github.com/PowerShell/openssh-portable).

Using `SIGINT`, it looks like the OpenSSH process doesn't do anything.

Unfortunately, it appears this causes OpenSSH to not cleanly disconnect
from the server.

This is not ideal, but it's probably acceptable for now since we'll stop
sending keep-alives and the connection should end up shutting down on
the server-side reasonably quickly (about 2 minute later).