Tunnel Helper: Force-kill after 5 seconds
This adds support for force-killing the SSH Helper after 5 seconds if it
doesn't exist immediately. This isn't required when running actual
OpenSSH, but when running the mock ssh.bat helper we use in tests, it
usually is (because sending `SIGBRK`, or even `SIGINT` to a console
doesn't always reliably terminate it).

NOTE: I'm leaving around a couple debug helpers in specs if needed in
the future.