Add a db:reload command
This adds a db:reload command, which recreates a DB container in-place
(same instance, same mountpoint).

The implementation is very similar to db:backup and I suspect we may
eventually want to merge the two, but considering there were a couple
subtle differences (the log message), I opted to not do so here.