New apps:scale option format
This changes the aptible apps:scale option format to treat container
count and container size as two "equal" options, rather than always
require container count to be provided as a positional argument and
allow container size to be provided as an afterthought.

To do so, both options are now positional arguments. I have however also
*renamed* one of the options: `--size` is now `--container-size`.

The rationale here is that we also use `--size` for `db:create` and
`backup:restore` in order to reference the size of the disk, but when we
introduce self-service DB scaling, we're going to want to also allow
specifying the container size, and that parameter will have to be named
differently (presumably `--container-size`).

In other words, for the sake of consistency, I'm updating `apps:scale`
to use `--container-` prefixes in the options so that the DB operations
can use the same parameters going forward.

Now, all these changes are potentially breaking changes, which we'd
rather avoid, so I've made sure to continue supporting the legacy
command line options (I did refactor most tests we had here, but I
confirmed the old tests were passing with the new code beforehand, and
I've kept tests for the old behavior around).

Finally, as we discussed, I also removed the size enum from the
`--container-size` option, since API will now be providing the list of
valid values if the user tries to use an unsupported size. This ensures
we won't need new CLI releases if new sizes are made available.