Add support for database scaling
This adds support for a `--container-size` option on `db:create` and
`db:reload`, and adds a new `db:restart` operation that can be used to
resize database containers and disks.

There isn't much going on here, with one notable exception: since we now
allow customers to pass an arbitrary size when creating databases via
`db:create`, we can easily get into a situation where we never attempt
provisioning a database, because the `provision` operation was rejected
in API.

In that case, we'll enqueue a deprovision operation to clean it up.