aptible ssh: return command exit code
This was an unfortunate regression introduced in #206: since we'd like
to clean up the certificate after ourselves (even though it'll expire
soon), we can no longer simply use Kernel::exec when running aptible
ssh.

I inadvertently converted this to use Kernel::system instead, but did
not pay attention to the fact that this swallows ssh's exit code. This
patch restores the original behavior. See the inline comments for how /
why it works.