config:X reject '-'-prefixed variable names
Variables with a leading '-' are likely to be mispelled arguments that
the user typed in: we've accepted them for 3 years, and there's only 3
apps with an option like this in their configuration, and they're all
mispelled.

So, let's just reject these options altogether.

This will be particularly useful for `aptible deploy`, where we'll be
allowing a number of actual options that could be misspelled with
confusing results (as opposed to the `config:X` commands, which only allow
`--app` and `--environment`, which will usually cause a fatal error if
missing).