Introduce aptible deploy
This command can be used in a few ways by customers:

First, it can be used to update an app to deploy from a Docker image (or
update its current Docker image), as well as set environment variables.

```
aptible deploy --app foobar --docker-image httpd:alpine FOO=BAR
```

When this is done, the app's `Dockerfile` will be ignored going forward.
However, other repo files might still influence the deploy.

Specifically, there are two possible cases:

- The app was deployed from git at some point in the past. In this case,
  the Docker image deploy will inherit the app's `Procfile` and
  `.aptible.yml`, if any. Enclave will fallback to the image's `CMD` as
  a `Procfile` substitute if none exists.
- The app was never deployed from `git`. In this case there is no
  `Procfile` or `.aptible.yml`, and Enclave uses the image's `CMD`. This
  is the recommended approach for users deploying from a Docker image.

It's also possible to completely remove the link to the git repo. When
that's done, the files found in the git repo will be ignored going
forward.

```
aptible deploy --app foobar --git-detach
```

But that's not all! It's also possible to *go back* to having a git repo
without a `git push`.

This is useful since a `git push` may not trigger a deploy if there are
no changes)

```
aptible deploy --app foobar --git-commitish master
```

Finally, it's possible to remove the app's Docker image to go back to a
`Dockerfile` build, albeit with a non-ideal syntax for now:

```
aptible deploy --app foobar APTIBLE_DOCKER_IMAGE=
```