Implement JSON output and a Logger
This refactors all our output to go through a Formatter that allows us
to prepare structured output to render as JSON or plain-text.

This allows us to always generate output identically regardless of the
output format (through the Formatter), and convert it to what the user
is asking for on the fly.

I've also introduced a logger while working on this, to ensure that
logging messages to to stderr, not stdout.