Create SSO "login" option (#259)
* Use provided tokens for SSO "login"

Part of https://aptible.atlassian.net/browse/DP-144

Allow the user to store a given token for later use to enable SSO
"logins" via the CLI. The SSO login process involves redirects that must
be done in a browser. Therefore, the way to allow users in SSO enforced
organizations to use the CLI is to provide a way for them to copy a
token from their browser (where they complete their SSO login) and store
it via the CLI for use.

The command is `aptible login --sso token'. If no token is provided, the
user will be prompted for one.

This persisted token will overwrite their existing one. Therefore, they
will need to issue `aptible login` again to access other orgs beyond the
one the SSO "organization" token is valid for.

* Bump version