With aptible login --sso, open browser for token.
When a user runs `aptible login --sso` with no token provided, we need
to prompt them for the token. And users will need to go to the
cli-sso-token page, so we may as well send them directly to the right
page.

This matches the behavior of similar commands like `aws sso login` from
the AWS CLI, which opens a browser window.