chore: upgrade aptible-api and aptible-auth gems, and allow minor version upgrades on each [SC-35150]
We're relaxing the aptible-cli gem dependencies to allow for minor version upgrades on aptible-auth and aptible-api. The specific changes to the gemspec are:

```
# original
  spec.add_dependency 'aptible-api', '~> 1.10.0'
  spec.add_dependency 'aptible-auth', '~> 1.2.5'

# new
  spec.add_dependency 'aptible-api', '~> 1.11'
  spec.add_dependency 'aptible-auth', '~> 1.4'
```

This _should_ be safe, because minor version upgrades in the semver framework include new features and maintain backwards compatibility.

The actual aptible-cli project, i.e., what we distribute and `bundle exec bin/aptible`, will continue to use specific versions of each of those gems based on the Gemfile.lock.